Complete assignment6

+++ server/controllers/filesController.js
@@ -25,6 +25,7 @@ const multer = require('multer');
 const multerS3 = require('multer-s3');
 const s3 = new aws.S3({ apiVersion: '2006-03-01' });
 const bucket = process.env.S3_BUCKET;
+// const bucket = "to run locally uncomment this!";
 const storage = multer.diskStorage({
     destination: (req, file, cb) => {
         cb(null, 'public/uploads')
@@ -35,15 +36,22 @@ const storage = multer.diskStorage({
 });
 
 const uploadS3 = multer({
+    fileFilter: function (req, file, callback) {
+        var ext = path.extname(file.originalname);
+        if (ext !== '.png' && ext !== '.jpg' && ext !== '.pdf' && ext !== '.jpeg') {
+            return callback({ "Error": "Only pdfs & images are allowed" }, false);
+        }
+        callback(null, true)
+    },
+    limits: {
+        fileSize: 1024 * 1024
+    },
     storage: multerS3({
         s3: s3,
         bucket: bucket,
         acl: 'private',
         contentType: multerS3.AUTO_CONTENT_TYPE,
         //   serverSideEncryption: 'AES256',
-        metadata: function (req, file, cb) {
-            cb(null, { fieldName: file.fieldname });
-        },
         key: function (req, file, cb) {
             cb(null, Date.now() + '-' + file.originalname)
         }
@@ -115,6 +123,7 @@ module.exports = {
                             } else {
                                 console.log("-----------------------------------------------------------------------")
                                 console.log(req.file);
+                                
                                 return File
                                     .create({
                                         id: uuidv4(),
@@ -124,7 +133,8 @@ module.exports = {
                                         size: req.file.size,
                                         fileOwner: user.dataValues.email_address,
                                         bill: bills[0].dataValues.id,
-                                        md5: req.file.metadata.fieldName
+                                        md5: "fieldName =>"+ req.file.fieldname,
+                                        key: req.file.key
                                     })
                                     .then((file) => {
                                         delete file.dataValues.createdAt;
@@ -133,6 +143,7 @@ module.exports = {
                                         delete file.dataValues.size;
                                         delete file.dataValues.bill;
                                         delete file.dataValues.md5;
+                                        delete file.dataValues.key;
                                         Bill
                                             .update(
                                                 { attachment: file.dataValues.id },
@@ -223,6 +234,7 @@ module.exports = {
                                 delete file[0].dataValues.size;
                                 delete file[0].dataValues.bill;
                                 delete file[0].dataValues.md5;
+                                delete file[0].dataValues.key;
                                 res.status(200).send(file[0]);
                             })
                             .catch((error) => {
@@ -309,23 +321,30 @@ module.exports = {
                                                 message: "File for this Bill Not Found!"
                                             })
                                         }
-
-                                        fs.unlink(file[0].dataValues.url, function (err) {
-
-                                            return File
-                                                .destroy({
-                                                    where: {
-                                                        id: req.params.fileId
-                                                    }
+                                        s3.deleteObject({
+                                            Bucket: bucket,
+                                            Key: file[0].key
+                                        }, function (err09) {
+                                            if (err09) {
+                                                return res.status(400).send({
+                                                    message: "Error while deleting from S3!"
                                                 })
-                                                .then((rowDeleted) => {
-                                                    if (rowDeleted === 1) {
-                                                        res.status(204).send('Deleted successfully');
-                                                    }
-                                                })
-                                                .catch((error2) => {
-                                                    res.status(400).send(error2);
-                                                });
+                                            } else {
+                                                return File
+                                                    .destroy({
+                                                        where: {
+                                                            id: req.params.fileId
+                                                        }
+                                                    })
+                                                    .then((rowDeleted) => {
+                                                        if (rowDeleted === 1) {
+                                                            res.status(204).send('Deleted successfully');
+                                                        }
+                                                    })
+                                                    .catch((error2) => {
+                                                        res.status(400).send(error2);
+                                                    });
+                                            }
                                         });
                                     })
                                     .catch((error) => {
